requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-environ==0.8.1
django-rest-authemail==2.1.2
django-oauth-toolkit==2.0.0
